Split shell project projection tests
